Fix autorepeat
xjump would leave X's autorepeat setting turned off if you closed it via
Alt+F4 or the close-window button.

I don't really know why it was turning off the autorepeat in the first place.
Everything seems to work fine when we leave autorepeat on all the time.

This is patch 2 of 5 in the official Debian package